Fixed issue with strtotime returning NULL with date formatted in a non-English locale
Fixed issue that caused dates which where regionally formatted as dd/MM/YYYY to be converted incorrectly. (For instance Sep. 27, 2018 was first being formatted as per the locale such that it was written 27/09/2018, and then was passed to strtotime() which explicitly expects either English formatted dates or a date in the ISO8601 format. Modified the code to pass the $mp_event_start_date into the call to strtotime() since $mp_event_start_date is holding the date in an ISO8601 format.