Modified date formatting code to consider WP Locale setting
The current code was making a call to the date() function to get an abbreviated month string.  This function explicitly returns the date in English.  Changed the code to use the datefmt_format_object method which allows specifying the locale to use for the formatting and updated the call to pass in the WP specified locale for the locale parameter.