Removed default value from end date check
In the 'if' statement that is checking for a non-existent end date the code was still providing a default value and so when no end date was found, there was still a default value coming back and since the 'if' was just checking for any value, the default value counted as any value and so end time was still being displayed.  Removed the default value so that if there is no end date then NULL is returned and the if statement now fails.